Make CurrencyEditText as private class
This is useful to avoid accidental uses outside of WCMaterialOutlinedCurrencyEditTextView
